Changes based on Adam's feedback: moved class, cleaned up scoping, readability
